migrate classes to the io.github.embeddedkafka.schemaregistry package
