--- kafka-streams/src/main/scala/net/manub/embeddedkafka/schemaregistry/streams/EmbeddedKafkaStreams.scala
@@ -1,16 +1,0 @@
-package net.manub.embeddedkafka.schemaregistry.streams
-
-import net.manub.embeddedkafka.schemaregistry.{
-  EmbeddedKafkaConfig,
-  EmbeddedKafka
-}
-import net.manub.embeddedkafka.streams.EmbeddedKafkaStreamsSupport
-
-trait EmbeddedKafkaStreams
-    extends EmbeddedKafkaStreamsSupport[EmbeddedKafkaConfig]
-    with EmbeddedKafka {
-  override protected[embeddedkafka] val streamsConfig =
-    new EmbeddedStreamsConfigImpl
-}
-
-object EmbeddedKafkaStreams extends EmbeddedKafkaStreams
